refactor(types): split up widget.ts
