+++ src/types/index.ts
@@ -14,3 +14,6 @@ export * from './insights';
 // widget-related
 export * from './connector';
 export * from './widget';
+export * from './ui-state';
+export * from './render-state';
+export * from './templates';
